+++ value.h
@@ -53,6 +53,7 @@ typedef struct {
 } RtnString;
 
 void ValueRelease(ValuePtr ptr);
+extern void* ValueToExternal(ValuePtr prt);
 extern RtnString ValueToString(ValuePtr ptr);
 const uint32_t* ValueToArrayIndex(ValuePtr ptr);
 int ValueToBoolean(ValuePtr ptr);
@@ -135,7 +136,7 @@ extern RtnValue NewValueBigIntFromWords(IsolatePtr iso_ptr,
 extern ValuePtr NewValueError(IsolatePtr iso_ptr,
                               ErrorTypeIndex idx,
                               const char* message);
-
+extern ValuePtr NewValueExternal(IsolatePtr iso_ptr, void* v);
 const char* ExceptionGetMessageString(ValuePtr ptr);
 
 extern void ObjectSet(ValuePtr ptr, const char* key, ValuePtr val_ptr);