+++ CHANGELOG.md
@@ -7,11 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 ### Added
+
 - Add support for ObjectTemplate.MarkAsUndetectable.
+- Add `Value.StrictEquals` providing strict equality checks in Go code
 
 ### Changed
-
-- Add `Value.StrictEquals` providing strict equality checks in Go code
 
 ## [v0.33.0] - 2025-05-15