+++ CHANGELOG.md
@@ -7,7 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 ### Added
+
 - Add support for ObjectTemplate.MarkAsUndetectable.
+- Add `Value.StrictEquals` providing strict equality checks in Go code
 
 ### Changed