+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add support for ObjectTemplate.MarkAsUndetectable.
 - Add `Value.StrictEquals` providing strict equality checks in Go code.
 - Add resource constraint options to `NewIsolate` in [#111](https://github.com/tommie/v8go/pull/111).
+- Add support for `Value.TypeOf()` corresponding to JavaScript `typeof` operator in [#104](https://github.com/tommie/v8go/pull/104).
 
 ### Changed